pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tasklings"
version = "0.1.0"
description = "Tasklings -- Family Chore Management PWA"
requires-python = ">=3.11"
dependencies = [
    "quart>=0.19",
    "aiosqlite>=0.20",
    "bcrypt>=4.1",
    "python-dotenv>=1.0",
    "itsdangerous>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
